将resourceManage基类和quotaManager、imageQuotaManager方法的调整
将quota相关的方法，从基类中放入到具体的业务相关的manager中，缩小方法可以调用的范围，只对相关的对象生效
