+++ src/main/java/com/yunionyun/mcp/mcclient/managers/impl/QuotaManager.java
@@ -1,5 +1,6 @@
 package com.yunionyun.mcp.mcclient.managers.impl;
 
+import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 import com.yunionyun.mcp.mcclient.EndpointType;
 import com.yunionyun.mcp.mcclient.JSONClientException;
@@ -46,4 +47,53 @@ public class QuotaManager extends ComputeManager {
         return this._get(s, url.toString(), this.keywordPlural);
     }
 
+    public JSONArray getV2(Session s, Map<String, String> param) throws McClientJavaBizException, IOException, JSONClientException {
+        StringBuilder url = new StringBuilder();
+        url.append("/");
+        url.append(this.keywordPlural);
+        if (param != null) {
+            Iterator var4 = param.entrySet().iterator();
+
+            while(var4.hasNext()) {
+                Map.Entry<String, String> entry = (Map.Entry)var4.next();
+                url.append("/" + (String)entry.getKey() + "/");
+                url.append((String)entry.getValue());
+            }
+        }
+
+        return this._getV2(s, url.toString(), this.keywordPlural);
+    }
+
+    public JSONArray addQuota(Session s, String type, String id, JSONObject body, boolean cascade) throws McClientJavaBizException, IOException, JSONClientException {
+        body.put("action", "add");
+        body.put("cascade", cascade);
+        return this.quotaSet(s, type, id, body);
+    }
+
+    public JSONArray resetQuota(Session s, String type, String id, JSONObject body, boolean cascade) throws McClientJavaBizException, IOException, JSONClientException {
+        body.put("action", "reset");
+        body.put("cascade", cascade);
+        return this.quotaSet(s, type, id, body);
+    }
+
+    private JSONArray quotaSet(Session s, String type, String id, JSONObject params) throws McClientJavaBizException, IOException, JSONClientException {
+        if (params == null) {
+            return null;
+        } else {
+            StringBuilder url = new StringBuilder();
+            url.append("/");
+            url.append(this.keywordPlural);
+            url.append("/");
+            url.append(type);
+            url.append("/");
+            url.append(id);
+            if ("domains".equals(type) && !params.containsKey("domain")) {
+                params.put("domain", id);
+            }
+
+            JSONObject body = new JSONObject();
+            body.put(this.keywordPlural, params);
+            return this._postV2(s, url.toString(), body, this.keywordPlural);
+        }
+    }
 }
